+++ crates/oxc_linter/src/rules/eslint/default_case_last.rs
@@ -59,7 +59,7 @@ impl Rule for DefaultCaseLast {
         let index_of_default = cases.iter().position(|c| c.test.is_none());
 
         if let Some(index) = index_of_default {
-            if index != cases.len() - 1 {
+            if index != cases.len() as usize - 1 {
                 let default_clause = &cases[index as u32];
                 ctx.diagnostic(default_case_last_diagnostic(Span::new(
                     default_clause.span.start,
